Add DbSeeder class for database seeding
Introduces the `DbSeeder` class to seed the database with initial data. Utilizes the `Bogus` library to generate fake data for events, sessions, session limits, users, and registrations. The `Seed` method checks for existing roles and entities, creating them if necessary, and includes private methods to generate realistic test data.

DbSeeder added to DI as Transistant and execute it in program.cs